@@ -20,20 +20,22 @@
     "dist/*.js"
   ],
   "scripts": {
-    "clean": "rimraf dist/",
-    "prepublish": "webpack",
-    "build": "webpack",
+    "clean": "rimraf dist/ && rimraf ../jupyter_widget_example/labextension/ && rimraf ../jupyter_widget_example/nbextension",
+    "prepublish": "yarn run clean && yarn run build:prod",
+    "build": "webpack --mode=development && yarn run build:labextension:dev",
+    "build:prod": "webpack --mode=production && yarn run build:labextension",
+    "build:labextension": "jupyter labextension build .",
+    "build:labextension:dev": "jupyter labextension build --development True .",
     "watch": "webpack --watch --mode=development",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "devDependencies": {
-    "webpack": "^3.5.5",
+    "@jupyterlab/builder": "^3.0.0",
+    "webpack": "^5",
     "rimraf": "^2.6.1"
   },
   "dependencies": {
-    "@jupyter-widgets/base": "^1 || ^2 || ^3",
-    "@jupyter-widgets/controls": "^1.1.5",
-    "lodash": "^4.17.4"
+    "@jupyter-widgets/base": "^1.1 || ^2 || ^3 || ^4 || ^6"
   },
   "jupyterlab": {
     "extension": "lib/labplugin"
