@@ -1,9 +1,4 @@
 # Module version
-version_info = (0, 1, 0, 'alpha', 0)
+__version__ = '0.1.0a0'
 
-# Module version stage suffix map
-_specifier_ = {'alpha': 'a', 'beta': 'b', 'candidate': 'rc', 'final': ''}
-
-# Module version accessible using jupyter_clickable_image_widget.__version__
-__version__ = '%s.%s.%s%s'%(version_info[0], version_info[1], version_info[2],
-  '' if version_info[3]=='final' else _specifier_[version_info[3]]+str(version_info[4]))
+NPM_PACKAGE_RANGE='^0.1.0'
