@@ -1,6 +1,6 @@
 import ipywidgets as widgets
 from traitlets import Unicode
-
+from ._version import NPM_PACKAGE_RANGE
 # See js/lib/example.js for the frontend counterpart to this file.
 
 @widgets.register
@@ -20,9 +20,9 @@ class ClickableImageWidget(widgets.Image):
     _model_module = Unicode('jupyter_clickable_image_widget').tag(sync=True)
 
     # Version of the front-end module containing widget view
-    _view_module_version = Unicode('^0.1.0').tag(sync=True)
+    _view_module_version = Unicode(NPM_PACKAGE_RANGE).tag(sync=True)
     # Version of the front-end module containing widget model
-    _model_module_version = Unicode('^0.1.0').tag(sync=True)
+    _model_module_version = Unicode(NPM_PACKAGE_RANGE).tag(sync=True)
 
     # Widget specific property.
     # Widget properties are defined as traitlets. Any property tagged with `sync=True`
